Some small changes

+++ src/protocols/dkg.rs
@@ -423,8 +423,8 @@ pub fn dkg_phase3(
         // We first compute a new session id.
         // As in Protocol 3.6 of DKLs23, we include the indexes from the parties.
         let mul_sid_receiver = [
-            &data.party_index.to_be_bytes(),
-            &i.to_be_bytes(),
+            &(data.party_index as u8).to_be_bytes(),
+            &(i as u8).to_be_bytes(),
             &data.session_id[..],
         ]
         .concat();
@@ -437,8 +437,8 @@ pub fn dkg_phase3(
         // New session id as above.
         // Note that the indexes are now in the opposite order.
         let mul_sid_sender = [
-            &i.to_be_bytes(),
-            &data.party_index.to_be_bytes(),
+            &(i as u8).to_be_bytes(),
+            &(data.party_index as u8).to_be_bytes(),
             &data.session_id[..],
         ]
         .concat();
@@ -610,8 +610,8 @@ pub fn dkg_phase4(
             // We retrieve the id used for multiplication. Note that the first party
             // is the receiver and the second, the sender.
             let mul_sid_receiver = [
-                &my_index.to_be_bytes(),
-                &their_index.to_be_bytes(),
+                &(my_index as u8).to_be_bytes(),
+                &(their_index as u8).to_be_bytes(),
                 &data.session_id[..],
             ]
             .concat();
@@ -640,8 +640,8 @@ pub fn dkg_phase4(
             // We retrieve the id used for multiplication. Note that the first party
             // is the receiver and the second, the sender.
             let mul_sid_sender = [
-                &their_index.to_be_bytes(),
-                &my_index.to_be_bytes(),
+                &(their_index as u8).to_be_bytes(),
+                &(my_index as u8).to_be_bytes(),
                 &data.session_id[..],
             ]
             .concat();
